Prevent including pre-compiled files
Currently it's not possible to compile @fluentui/react package with any style plugin. From the esbuild thread https://github.com/evanw/esbuild/issues/1834 it is mentioned the plugins should take care of properly selecting the files to compile. Thus I added 3 simple lines of code which will prevent inclusion of those precompiled files. In the example of fluentui there is for instance the import for BasePicker.less which indeed exists but it will instead picked the pre-compiled BasePicker.less.js.
Every esbuild style plugin tries to compile these JS files with less or what ever. So the condition just makes sure the resolved file is really a style.